Added a check and confirmation for overwriting the systemctl policy

+++ setup.sh
@@ -29,7 +29,20 @@ for module in $modules ; do
 	./testmod ${module} || abort "Cannot import ${module} - this is required for installation"
 done
 echo "Modules are ok, installing files"
-install --backup -m 644 ${policy} ${policy_path}/${policy} || abort "Error installing policy"
+if [[ -f "${policy_path}/${policy}" ]] ; then
+	while true ; do
+		read -p ""${policy_path}/${policy}" exists, overwrite? (it will be backed up) [Yes/no]: " answer
+		answer=${answer:=yes}
+		if [[ "$answer" == "no" ]] ; then
+			echo "${policy_path}/${policy} not overwritten, this might not work"
+			break
+		elif [[ "$answer" == "Yes" ]] || [[ "$answer" == "yes" ]] ; then
+			install --backup -m 644 ${policy} ${policy_path}/${policy} || abort "Error installing policy"
+			break
+		fi
+		echo "You must answer yes or no"
+	done
+fi
 install --backup -D -m 644 ${rules} ${rules_path}/${rules} || abort "Error installing rules"
 install --backup -m 755 sview ${binloc}/sview || abort "Error installing program"
 xdg-desktop-menu install --novendor ${desktop}
